fastapi/blogapi/pyproject.toml
[tool.poetry]
name = "fastapi-boilerplate"
version = "0.1.1"
description = ""
authors = ["Tarang Ambaliya <taranga@webelight.co.in>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.88.0"
celery = "^5.2.7"
pyjwt = "^2.6.0"
sqlalchemy = "^1.4.42"
alembic = "^1.8.1"
uvicorn = "^0.19.0"
aioredis = "^2.0.1"
ujson = "^5.5.0"
fastapi-utils = "^0.2.1"
pydantic = {extras = ["dotenv", "email"], version = "^1.10.2"}
asyncpg = "^0.27.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
